Updated past 2 weeks of work
Went through multiple iterations of the controller input -> motor output. First was a switch on/off where anything over 15 would translate to full power in that direction. Next, we did a linear approach where the controller input (-127 -> 127) was directly applied to the voltage (also -127 -> 127). Finally, landed on this current version that uses a smoothing function to map the -127 -> 127 to an exponential curve giving smaller more precise control  at lower input still with the ability to go full speed.